requirements.txt
playwright==1.53.0
fake-useragent==2.2.0
python-dotenv==1.0.1
fastapi==0.111.0
uvicorn[standard]==0.30.1